HvD: Updating the QA run mechanisms to reduce the reliance on .nwchemrc as much as possible. The QA is most often run on platforms where people have multiple versions of NWChem. Sometimes people might even be testing different versions simultaneously. The .nwchemrc mechanism is rather unflexible in those cases and using environment variables would probably be ideal. However, at present not all settings can be controlled through environment variables. Hence the .nwchemrc still is needed.
Previously we tried do deal with .nwchemrc by moving an existing one out of
the way, and writting one specific for the tests, running the tests and then
moving the old one back. This procedure only allows running one set of tests
at a time. In case two tests are simultaneously the set that starts second would
detect that a back-up of .nwchemrc already exists and abort. It would be quite
tedious to write something that would guarantee that at the end of any number
of simultaneous tests the original .nwchemrc would be restored (the cost of
doing this would almost certainly outweigh the benefit). Hence this mechanism
is strictly limiting us to running one set of tests at a time per user.

To eliminate this problem the approach has been changed in the following way:
If no .nwchemrc is found or if the .nwchemrc file is writable then whenever the
tests start they will overwrite the existing .nwchemrc file. If there is
a readonly .nwchemrc file then the test cases will now use that file and rely
on its contents being valid. The .nwchemrc file only contains the setting for
the MD force fields as all other variables can be controlled through 
environment variables which is the preferred mechanism for the QA. Users may
still add other settings to the .nwchemrc file as the environment variables
take precedence over the .nwchemrc file. They need to remember though to 
write-protect their .nwchemrc file to avoid losing its contents.


git-svn-id: https://svn.pnl.gov/svn/nwchem/trunk@22633 85dd5168-0e29-45a6-bdf9-2122c44e486f