HvD: COSMO analytical gradients added. Because of the fact that the COSMO charges move along with the atoms some new derivative integrals were needed. I.e. for the interaction of a shell-pair with a COSMO charge we need the derivatives with respect to the basis function centers as well as the COSMO charge position. This way we can effect the proper cancellations and additions of contributions when the forces on the COSMO charges are summed onto the forces of the atoms.
At the moment the integral routines I have created are rather simplistic in
that they treat one COSMO charge at a time. There is some potential for 
optimization there.

Another thing is that the default settings for the number of COSMO charges per
atom is too crude for geometry optimizations. Because a COSMO charge is 
eliminated as soon as its center of charges moves into the radius around another
atom there are discontinuities in the potential energy surface. The size of 
these discontinuities depends on the surface area associated with a COSMO
charge. So with few charges per atom the potential energy surface becomes so
noisy that the geometry optimization never converges (irrespective of 
numerical or analytical gradient based optimization algorithms). So far my
experience shows that minbem=3 and maxbem=5 are the minimum settings needed
for geometry optimizations.

Finally, at present the linear system solver for the COSMO charges is still
a serial solver. This causes the COSMO charge evaluation to be disproportionally
expensive in parallel calculations.

Test cases for this functionality are to follow.


git-svn-id: https://svn.pnl.gov/svn/nwchem/trunk@22732 85dd5168-0e29-45a6-bdf9-2122c44e486f