HvD: Introduced a bit of a hack to make NWChem smarter about finding ECPs. Previously, if a user specified a basis set and used the same name for the ECP then NWChem would find the ECP only if it was present in the basis set file with that name. For many ECPs that is not the case as there is another ECP associated with the particular basis set. This is even more confusing because the Basis Set Exchange will automatically resolve the ECP from the association. So a user getting the basis set from the BSE will get something including an ECP whereas NWChem will complain it cannot find the ECP.
The resolution of this issue is to have NWChem try first to find the 
ECP in current basis set file. If that fails the code will now look for
the ASSOCIATED_ECP keyword and pick the ECP basis set name up from that.
Then the code will search for the specified ECP again (and hopefully
find it). The hope is that this makes dealing with ECPs a bit more 
transparent to users.


git-svn-id: https://svn.pnl.gov/svn/nwchem/trunk@25852 85dd5168-0e29-45a6-bdf9-2122c44e486f