HvD: The density functionals with derivatives up to and including third order generated by automatic differentiation using the univariate Taylor series approach by Griewank et al. DOI: 10.1090/S0025-5718-00-01120-0.
For this code to perform well it is essential that the compiler inlines a lot
of code. If the overloaded operators are evaluated as function calls the 
calculations slow down by more than an order of magnitude. Some compilers
can only inline code within a single file. Hence the addition of nwxc.F
that includes all the source code of the various functionals. In particular
the GCC generated code saw significant performance improvements as a result.
For GCC one is adviced to use compiler versions post 4.6 as apparently 
significant improvements to the inline capabilities were introduced at that
point.

At this point the gradient and 2nd derivative evaluation have been tested.
For the 3rd order derivatives test cases that actually exploit those derivatives
are still needed. If there are any deficiencies at that level they are most
likely to stem from the driver routine NWXC_EVAL_DF3_DRIVER which has to
generate the appropriate partial derivatives and interpolate the final results.
The generation of the derivatives themselves should work alright as the unit
tests in src/nwxc/nwad/unit_tests work and the lower order order derivatives
of the functionals work.

At the moment the performance is still not as good as I would like it. I suspect
that the univariate approach is in part to blame for that. The fact that all
lower order derivatives need to be re-evaluated for every highest order partial
derivative introduces an overhead that is close to proportional to the
highest order of derivative requested. I am planning to try a multivariate 
approach instead. This however has increased memory access as a downside,
in particular during assignments. We will see what is best on balance.


git-svn-id: https://svn.pnl.gov/svn/nwchem/trunk@26056 85dd5168-0e29-45a6-bdf9-2122c44e486f