HvD: Eric found that the Fortran 2008 intrinsics POPCNT, LEADZ and TRAILZ are not supported by older compilers. This caused problems building the NWXC libraries on a variety of machines. Hence he added regular Fortran implementations of these functions. I have tweaked these implementations a little bit using BTEST (Fortran 1995) which seems to work on with all compilers tried so far. I have also added a little description.
Having explicit implementations of these functions did have a knock
on effect on the NWAD implementation. This module actually uses these
functions to keep track of the independent variables a given quantity
depends on (easy to do because the number of variables a functional
depends on is maximally 9 in practice). To save memory I had explicitly
declared the corresponding variable as INTEGER*4. However, NWChem 
typically promotes all integers to INTEGER*8 causing the Fortran 
implementation of LEADZ and TRAILZ to clash with the explicit 
INTEGER*4 data type. Hence I have removed the explicit INTEGER*4 
declaration from the module. The NWXC library is written in such a 
way as to automatically detect the size of the TYPE(NWAD_DBLE) 
data structures so no changes are needed to account for that.


git-svn-id: https://svn.pnl.gov/svn/nwchem/trunk@26354 85dd5168-0e29-45a6-bdf9-2122c44e486f