HvD: A problem with (some) Meta-GGAs was uncovered. The problem is that the kinetic energy density can be zero at points where the density is not zero (e.g. the mid-point in H2). At such points the derivatives of the M06 functionals are singular. Hence if you are unlucky enough to pick a grid point close to such a point your calculation might blow up.
This problem was analyzed by Grafenstein et al. in J.Chem.Phys. 127,
214103 (2007), doi: 10.1063/1.2800011. They identified the problem but 
their analysis and solution are shaky (I guess they wrote the paper in
a hurry). Their solution for example fixes the first order derivatives
but the problem remains in higher order derivatives.

The solution we use at the moment is to approximate the gradient and
kinetic energy density dependent enhancement factor with 1 if the
kinetic energy density approaches 0. In effect that means we replace
the meta-GGA with LDA at that point (we have to keep the LDA part as
the density is non-zero and therefore there is a contribution to the
energy).

The problem surfaced in studies of nitroaniline during the geometry
optimization and subsequent numerical frequency evaluation. The
geometry optimization would typically converge despite occasional bad
gradients. The frequency calculations would produce crazy results with
frequencies like positive and negative 100,000s.


git-svn-id: https://svn.pnl.gov/svn/nwchem/trunk@26772 85dd5168-0e29-45a6-bdf9-2122c44e486f