HvD: Removed the use of chi.
A number of functionals remain that explicitly use chi. These
functionals include:
- Becke 1988 (and of course CAM-B88)
- HSE
- Gill 1996
- Omege PBE
- Perdew - Wang 1991
- Perdew - Wang 2006
For these functionals it is still possible to replace all even powers
of chi by equations that are directly expressed in terms of gamma
(except for Gill 1996 which does not include any even powers of chi).
However, there are also odd powers of chi for which nothing can be done
to mitigate problems with the derivatives when gamma-->0. Hence at this
point in time I am not going to deal with this (and hopefully we will
never HAVE to deal with this).


git-svn-id: https://svn.pnl.gov/svn/nwchem/trunk@26838 85dd5168-0e29-45a6-bdf9-2122c44e486f