HvD: Tightening up the screening on small densities. In automatic differentiation an assignment does two things: 1. It gives a value to a variable 2. It sets the data dependencies of a variable. These two things are not the same. Hence assigning rhob=rho(n,R_B) when the beta density is tiny is not the same as setting rhob=0.0d0. The value does not differ much in both cases but the data dependencies do. In the first case rhob depends on the beta density and the corresponding derivatives will be calculated, in the second case rhob is just a constant and no derivatives will be calculated.
In itself this does not lead to differences in the results, but when
nwxc_fnc_print prints the energy for the fully spin-polarized case it changes
the expression generated. In particular the expressions included references
to rhob, sigmaab, and sigmabb that Maxima did not expect. Hence uninitialized
variables were introduced.

Now the screening has been adapted so that when inputs have values below
the tolerance the variables internally get set to zero. This eliminates the
spurious data dependencies.


git-svn-id: https://svn.pnl.gov/svn/nwchem/trunk@27111 85dd5168-0e29-45a6-bdf9-2122c44e486f