HvD: Python has developed a mechanism that allows multiple versions and multiple variants of a single version to be installed at the same time.
Part of this change is the introduction of an ABI (Application Build
Interface?). The ABI is indicate by ABI flags of which there currently are
three: d - debug, m - memory allocator, u - unicode, leading to different
possible combination from nothing to dmu. 

A result of this change is that a lot more things are now also labeled with
version numbers and ABI flags, such as config scripts, library files, 
directories, etc. Ultimately we will have to figure out when all this stuff 
applies (i.e. from which versions of Python) and to what entities it applies
how. We will this to ensure we can build NWChem with Python support
successfully. 

This commit makes a start by introducing a mechanism to establish whether
the current version of Python supports ABI flags and if so what they are.
From here we can propagate this ABI stuff further throughout the build system
as we figure more parts of this puzzle out.


git-svn-id: https://svn.pnl.gov/svn/nwchem/trunk@27134 85dd5168-0e29-45a6-bdf9-2122c44e486f