eliminate absmax reductions in ga_[id]gop calls
the semantics of absmax and absmin reductions in GA/ARMCI
are gross, as the input buffer is overwritten, even if
the operation is executed by only process.  this is
inconsistent with MPI collectives and user expectations.

by making the absolute value application explicit in the application,
we eliminate any ambiguity in the code and allow the GA/ARMCI/MPI
runtime library to support collectives in a more natural way.



git-svn-id: https://svn.pnl.gov/svn/nwchem/trunk@27936 85dd5168-0e29-45a6-bdf9-2122c44e486f